move game functionality into game.js, refactor imports and functions as necessary
